+++ src/main/java/domain/participant/Participant.java
@@ -1,7 +1,9 @@
 package domain.participant;
 
 import domain.card.Card;
-import domain.card.Hand;
+import domain.participant.info.Betting;
+import domain.participant.info.Hand;
+import domain.participant.info.Name;
 
 public class Participant {
