+++ src/main/java/domain/participant/info/Betting.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.participant.info;
 
 public class Betting {
 	private static final String BETTING_MONEY_NEGATIVE_ERROR_MESSAGE = "[Error] 배팅 금약은 0원 이상이어야 합니다.";
