+++ src/main/java/view/InputView.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
-import domain.participant.Name;
+import domain.participant.info.Name;
 
 public class InputView {
