+++ src/test/java/domain/participant/DealerTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 
 import domain.card.Card;
 import domain.card.Denomination;
-import domain.card.Hand;
 import domain.card.Suit;
+import domain.participant.info.Hand;
 
 public class DealerTest {
