+++ src/test/java/domain/result/ResultTest.java
@@ -12,13 +12,13 @@ import org.junit.jupiter.api.Test;
 
 import domain.card.Card;
 import domain.card.Denomination;
-import domain.card.Hand;
 import domain.card.Suit;
-import domain.participant.Betting;
 import domain.participant.Dealer;
-import domain.participant.Name;
 import domain.participant.Player;
 import domain.participant.Players;
+import domain.participant.info.Betting;
+import domain.participant.info.Hand;
+import domain.participant.info.Name;
 
 public class ResultTest {
